tipoventa

+++ src/main/java/com/pedidos/mayorista/model/DetallePedido.java
@@ -21,7 +21,7 @@ public class DetallePedido {
     @ManyToOne
     private Producto producto;
 
-    private Integer cantidad;
+    private BigDecimal cantidad;
 
     private BigDecimal subtotal;
 }
